CSV und PDF für Buchungen und Anlagenbuchungen
