+++ src/de/jost_net/JVerein/keys/VorlageTyp.java
@@ -122,6 +122,8 @@ public enum VorlageTyp
       "Summenbuchungen Dateiname",
       "Summenliste-$filter_datum_von_f-$filter_datum_bis_f-$aktuellesdatum-$aktuellezeit",
       Vorlageart.DATEINAME.getKey()),
+  BUCHUNGEN_DATEINAME("buchungen-dateiname", "Buchungen Dateiname",
+      "Buchungen-$aktuellesdatum-$aktuellezeit", Vorlageart.DATEINAME.getKey()),
   ANLAGEN_BUCHUNGSJOURNAL_DATEINAME("anlagen-buchungsjournal-dateiname",
       "Anlagen Buchungsjournal Dateiname",
       "Anlagen-Buchungsjournal-$filter_datum_von_f-$filter_datum_bis_f-$aktuellesdatum-$aktuellezeit",
@@ -137,6 +139,10 @@ public enum VorlageTyp
   ANLAGEN_SUMMENBUCHUNGEN_DATEINAME("anlagen-summenbuchungen-dateiname",
       "Anlagen Summenbuchungen Dateiname",
       "Anlagen-Summenliste-$filter_datum_von_f-$filter_datum_bis_f-$aktuellesdatum-$aktuellezeit",
+      Vorlageart.DATEINAME.getKey()),
+  ANLAGEN_BUCHUNGEN_DATEINAME("anlagen-buchungen-dateiname",
+      "Anlagen Buchungen Dateiname",
+      "Anlagen-Buchungen-$aktuellesdatum-$aktuellezeit",
       Vorlageart.DATEINAME.getKey()),
   BUCHUNGSKLASSENSALDO_DATEINAME("buchungsklassensaldo-dateiname",
       "Buchungsklassensaldo Dateiname",
@@ -350,6 +356,10 @@ public enum VorlageTyp
       "Summenbuchungen Subtitel",
       "$udateformat.format(\"dd.MM.yyyy\",$udateformat.parse(\"yyyyMMdd\",$filter_datum_von_f)) - $udateformat.format(\"dd.MM.yyyy\",$udateformat.parse(\"yyyyMMdd\",$filter_datum_bis_f))",
       Vorlageart.TITEL.getKey()),
+  BUCHUNGEN_TITEL("buchungen-titel", "Buchungen Titel", "Buchungen",
+      Vorlageart.TITEL.getKey()),
+  BUCHUNGEN_SUBTITEL("buchungen-subtitel", "Buchungen Subtitel", "",
+      Vorlageart.TITEL.getKey()),
   ANLAGEN_BUCHUNGSJOURNAL_TITEL("anlagen-buchungsjournal-titel",
       "Anlagen Buchungsjournal Titel", "Buchungsjournal",
       Vorlageart.TITEL.getKey()),
@@ -371,6 +381,10 @@ public enum VorlageTyp
       "Anlagen Summenbuchungen Subtitel",
       "$udateformat.format(\"dd.MM.yyyy\",$udateformat.parse(\"yyyyMMdd\",$filter_datum_von_f)) - $udateformat.format(\"dd.MM.yyyy\",$udateformat.parse(\"yyyyMMdd\",$filter_datum_bis_f))",
       Vorlageart.TITEL.getKey()),
+  ANLAGEN_BUCHUNGEN_TITEL("anlagen-buchungen-titel", "Anlagen Buchungen Titel",
+      "Buchungen", Vorlageart.TITEL.getKey()),
+  ANLAGEN_BUCHUNGEN_SUBTITEL("anlagen-buchungen-subtitel",
+      "Anlagen Buchungen Subtitel", "", Vorlageart.TITEL.getKey()),
   BUCHUNGSKLASSENSALDO_TITEL("buchungsklassensaldo-titel",
       "Buchungsklassensaldo Titel", "Buchungsklassensaldo",
       Vorlageart.TITEL.getKey()),
